requirements.txt
aioredis==2.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio
async-timeout
bidict
certifi
cffi
click
colorama
cryptography
dnspython
fastapi
greenlet
h11
httpcore
httpx
idna
motor
passlib
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT
pymongo
python-dotenv
python-engineio
python-multipart
python-socketio
redis
simple-websocket
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn
wsproto
supabase
